Implement Vietoris-Rips algorithm enhancements and fix warnings
- Added a lightweight reminder script and scheduled it to run.
- Created a minimal Vietoris-Rips unit test to assert non-empty simplices and reasonable Betti numbers.
- Cleaned up warnings in DTM filtration and vectorization code, addressing narrowing conversions and unused variables.
- Updated DTMFiltration to ensure simplex tree existence and handle vertex indexing safely.
- Enhanced VietorisRipsImpl initialization with input validation.
- Modified vectorization classes to ensure proper type handling for dimensions.
- Re-enabled and updated the test_vietoris_rips.cpp file to use standard assertions instead of GoogleTest framework.
- Improved error handling and added tests for move semantics and larger point clouds.
- Ensured all tests pass successfully with appropriate assertions.

+++ src/cpp/vectorization/betti_curve.cpp
@@ -100,7 +100,7 @@ int BettiCurve::calculateBettiNumber(
     // Count how many homology features of the given dimension are active at this filtration value
     int count = 0;
     for (const auto& pair : diagram) {
-        if (pair.dimension != dimension) continue;
+        if (static_cast<int>(pair.dimension) != dimension) continue;
         
         // A feature is active if it has been born but not died yet
         if (pair.birth <= filtration_value && filtration_value < pair.death) {
